Cleanup check for openGL errors in Android driver.
Use an inline function instead of a macro as there is no point in using
a macro here beside inlining code. Rename GL_CHECK_ERR to gl_check_error
to respect coding style for functions. It makes clear that GL_CHECK_ERR
is actually doing something by having the () next to the call. It fixes
up indentation by forcing the need for a semicolon after the call.