Install app icon in XDG hicolor icon theme
Install the icon of the application in the hicolor XDG icon theme;
this way it can be properly loaded by XDG menus in the currently set
XDG icon theme, without looking in the legacy pixmaps location.
Because of this, there is no more need to specify the 'Icon' key in
the desktop file as absolute path, as an icon name is searched in the
XDG icon theme.

Adapt also the Debian install file for the changed location.